pyproject.toml
[tool.poetry]
name = "django-resource-viewset"
version = "0.1.0"
description = ""
authors = ["Chris Muthig <camuthig@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
